Filter a dataset

A function to filter a data frame based on specified list. The list should contain elements with the following structure: - col: the name of the column to filter on - fun: the filtering function to use ("between" or "in") - min, max: the range for "between" filters - val: the values for "in" filters

Modified from Teofil Nakov's blog: https://discindo.org/posts/2025-01-31-filter-snippet/

.filter(data, arglist)

Arguments

data

the data frame

arglist

a list of filter specifications

Examples

d <- iris
l <- list(
  list(col = "Sepal.Length", fun = "between", min = 4, max = 5),
  list(col = "Species", fun = "in", val = c("setosa", "versicolor"))
)
.filter(d, l)
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1           4.9         3.0          1.4         0.2     setosa
#> 2           4.7         3.2          1.3         0.2     setosa
#> 3           4.6         3.1          1.5         0.2     setosa
#> 4           5.0         3.6          1.4         0.2     setosa
#> 5           4.6         3.4          1.4         0.3     setosa
#> 6           5.0         3.4          1.5         0.2     setosa
#> 7           4.4         2.9          1.4         0.2     setosa
#> 8           4.9         3.1          1.5         0.1     setosa
#> 9           4.8         3.4          1.6         0.2     setosa
#> 10          4.8         3.0          1.4         0.1     setosa
#> 11          4.3         3.0          1.1         0.1     setosa
#> 12          4.6         3.6          1.0         0.2     setosa
#> 13          4.8         3.4          1.9         0.2     setosa
#> 14          5.0         3.0          1.6         0.2     setosa
#> 15          5.0         3.4          1.6         0.4     setosa
#> 16          4.7         3.2          1.6         0.2     setosa
#> 17          4.8         3.1          1.6         0.2     setosa
#> 18          4.9         3.1          1.5         0.2     setosa
#> 19          5.0         3.2          1.2         0.2     setosa
#> 20          4.9         3.6          1.4         0.1     setosa
#> 21          4.4         3.0          1.3         0.2     setosa
#> 22          5.0         3.5          1.3         0.3     setosa
#> 23          4.5         2.3          1.3         0.3     setosa
#> 24          4.4         3.2          1.3         0.2     setosa
#> 25          5.0         3.5          1.6         0.6     setosa
#> 26          4.8         3.0          1.4         0.3     setosa
#> 27          4.6         3.2          1.4         0.2     setosa
#> 28          5.0         3.3          1.4         0.2     setosa
#> 29          4.9         2.4          3.3         1.0 versicolor
#> 30          5.0         2.0          3.5         1.0 versicolor
#> 31          5.0         2.3          3.3         1.0 versicolor